Improvements to auditing and improvements to the Step 3.2.6/3.2.7 function output filing
Improvements to auditing - corrections to names and locations of some audit files, corrections to content of others. All auditing inprovements post 3.2.5 (i.e. 3.2.6 and onward). Also added the 3.2.6 output to the files dumped from the Step 3.2.6/3.2.7 function, not just the 3.2.7 output! Also includes an extra script that can be used to extract a given sector from the large compressed output files and dump it into a nice readable csv for checking purposes.