Reformatting input data to a DVector-readble format.
Reformat_input_data.py reads in each data source, does some processing to get into a DVector friendly format (the aim is to genericise this as much as poss), and then export an HDF5 format file in the same location as the input file (in a preprocessing subfolder) with the same file name. This *should* help the reader to know which file is being used.