Custom segmentations standardised and generated based on yaml.
Segmentations provided by the yaml file are checked against the caf.core SuperSegments to see if they already exist, and then the segmentations are generated based on this combination of standard and custom segmentations.

Raises an error if the segmentations provided in the yaml are not defined either in the constants.segments.py, or in the caf.core SuperSegments.