Creation of Output 2.
Added the missing 'h' category to the MSOA level data from ons table 2 and infilled with zeros. This is done in parsers.py and needs genericising. Maybe look for missing segments from the core definition and infill?

Need to control the output 2 to the 2022 population distribution by LSOA before moving on to Output 3.